fix single file

+++ src/app/Library/Uploaders/SingleFile.php
@@ -54,12 +54,17 @@ class SingleFile extends Uploader
         }
 
         foreach ($previousRepeatableValues as $row => $file) {
-            if ($file) {
-                if (! isset($orderedFiles[$row])) {
-                    $orderedFiles[$row] = null;
+            foreach ($previousRepeatableValues as $row => $file) {
+                if ($file) {
+                    if(! isset($orderedFiles[$row])) {
+                        $orderedFiles[$row] = null;
+                    
+                    }
+                    if($file !== $orderedFiles[$row]) {
+                        Storage::disk($this->getDisk())->delete($file);
+                    }
+                
                 }
-
-                Storage::disk($this->getDisk())->delete($file);
             }
         }
